+++ src/main/java/groupId/artifactId/service/api/IMenuService.java
@@ -2,7 +2,7 @@ package groupId.artifactId.service.api;
 
 import groupId.artifactId.core.dto.MenuItemDto;
 import groupId.artifactId.core.dto.MenuItemDtoWithId;
-import groupId.artifactId.storage.entity.api.IMenu;
+import groupId.artifactId.dao.entity.api.IMenu;
 
 import java.util.List;
 
@@ -10,6 +10,6 @@ public interface IMenuService {
     void add(List<MenuItemDto> menuItemDto);
     void addMenuItem(MenuItemDtoWithId menuItemDtoWithId);
     List<IMenu> get();
-    Boolean isIdValid(int id);
+    Boolean isIdValid(Long id);
     Boolean isDishExist(String name);
 }
